+++ user-service/src/main/java/com/study/userservice/config/MapperConfig.java
@@ -1,4 +1,4 @@
-package com.study.userservice.configs;
+package com.study.userservice.config;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
