--- user-service/src/main/java/com/study/userservice/dto/request/CreateUserRequest.java
@@ -1,33 +1,0 @@
-package com.study.userservice.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.study.userservice.enums.Gender;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Past;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateUserRequest {
-    @NotBlank(message = "Username is required.")
-    @Size(min = 3, max = 20, message = "Username must be between 3 and 20 characters.")
-    private String username;
-
-    @NotNull(message = "Date of birth is required.")
-    @Past(message = "Date of birth must be in the past")
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate dateOfBirth;
-
-    @NotNull(message = "Gender is required.")
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private Gender gender;
-}
