--- user-service/src/main/java/com/study/userservice/dto/request/UpdateUserRequest.java
@@ -1,30 +1,0 @@
-package com.study.userservice.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.study.userservice.enums.Gender;
-import jakarta.validation.constraints.Past;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateUserRequest {
-    @Size(min =3, max = 20, message = "Username must be between 3 and 20 characters.")
-    private String username;
-
-    @Past(message = "Date of birth must be in the past")
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate dateOfBirth;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private Gender gender;
-
-    private String avatarUrl;
-}
