--- user-service/src/main/java/com/study/userservice/dto/response/UserResponse.java
@@ -1,27 +1,0 @@
-package com.study.userservice.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-import java.util.UUID;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserResponse {
-    private UUID id;
-
-    private String username;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate dateOfBirth;
-
-    private String gender;
-
-    private String avatarUrl;
-}
