--- user-service/src/main/java/com/study/userservice/exceptions/BusinessException.java
@@ -1,7 +1,0 @@
-package com.study.userservice.exceptions;
-
-public class BusinessException extends RuntimeException {
-    public BusinessException(String message) {
-        super(message);
-    }
-}
