--- user-service/src/main/java/com/study/userservice/exceptions/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.userservice.exceptions;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
