+++ user-service/src/main/java/com/study/userservice/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.study.userservice.repositories;
+package com.study.userservice.repository;
 
-import com.study.userservice.models.User;
+import com.study.userservice.enity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
